Lift Gravatar component into @automattic/jetpack-components
Move Forms' `Gravatar` (profile avatar with hovercard) into
`@automattic/jetpack-components` so Forms and the Subscribers Dashboard
can share one implementation. Subscribers had a near-identical local
copy from the umbrella PR (#48386); both now import the shared one
from the package's new `./gravatar` subpath export, which keeps esbuild
out of the rest of the components tree (subpath import avoids loading
the whole index, which triggers SCSS resolution failures under wp-build).

- New `components/gravatar/{index.tsx,style.scss}` in `js-packages/components`,
  plus a `Gravatar` named export and a `./gravatar` subpath export.
- `js-sha256` and `@gravatar-com/hovercards` move with the component;
  Forms drops them from its top-level dependencies.
- Forms call sites (`routes/responses/stage.tsx`, the inbox stage,
  and `inspector/response-meta`) now import from the shared package.
- Subscribers Dashboard drops `_inc/components/gravatar.tsx` and the
  duplicated dependencies from both the package and the route manifest.

Long term, the shared component is the seam for swapping in core's
Gravatar / Hovercard primitive once that lands
(see https://github.com/WordPress/gutenberg/issues/76836).
